import-bundle: move to SES-0.8.0, abandon non-ses mode
