+++ packages/import-bundle/package.json
@@ -13,13 +13,10 @@
     "lint-fix": "eslint --fix '**/*.js'",
     "lint-check": "eslint '**/*.js'"
   },
-  "dependencies": {
-    "ses-adapter": "^0.0.6"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@agoric/bundle-source": "^1.1.4",
+    "@agoric/install-ses": "^0.1.0",
     "esm": "^3.2.5",
-    "ses": "^0.7.7",
     "tap": "^14.10.5"
   },
   "files": [
