+++ packages/import-bundle/test/bundle2.js
@@ -1,3 +1,7 @@
-export function bundle2(a) {
+export function bundle2_add(a) {
   return a + 2;
+}
+
+export function bundle2_transform(a) {
+  return `${a} is two foot wide`;
 }
