--- packages/import-bundle/test/test-non-ses.js
@@ -1,2 +1,0 @@
-// disabled because our non-SES 'Compartment' is still a non-functioning stub
-// import './do-import.js';
